859: Renamed TransactionLink to AccountLink
